tests/requirements-tests.txt
# recommended using python >= 3.8

numpy~=1.23.3
pandas~=1.5.0
scikit-learn~=1.1.2
mlflow~=2.4.1
minio~=7.1.11
kfp~=1.8.10
boto3~=1.24.80
pytest~=7.1.3
pytest-order~=1.0.1
pytest-timeout==2.1.0
jq~=1.3.0
python-dotenv~=0.21.0
envsubst~=0.1.5
